Streaming of song playback info - total, elapsed, remaining time
The server sends an update to the client in intervals of (by default)
1 second.
The update contains:
- The duration of the song, which can sometimes be inaccurate when
  mpv is unable to obtain the data, e.g. when streaming mp3 files.
- The current playback time.
- The time remaining until the end of the song. It can be inaccurate
  as well since it is dependant on the duration of the song. It is
  basically duration - current_time.
- The adjusted remaining time (called remaining playtime). It is
  adjusted by the playback speed (which cannot be yet controlled by
  this program).
- The current playback state: PLAYING, PAUSED, STOPPED, FINISHED or
  FINISHED_ERROR. FINISHED and FINISHED_ERROR will never be returned
  by PlaybackState::GetState(), these states are there only to fire a
  callback when the song has finished playing on its own.
The update interval is ignored whenever the playback state changes.
In that case the update is sent immediately.

For now, the daemon doesn't do anything with this information, only
displays it when in the debug mode.
It's designed to be a synchronization mechanism with a music player.
This feature is enabled by default, for now there is no way to disable
it.