Replace SongInfoUpdater with PlaybackSynchronizer
PlaybackSynchronizer is a class that does the same stuff as
SongInfoUpdater but with better name and better encapsulation.
It stores the whole playback information that can be obtained by
calling PlaybackSynchronizer::GetPlaybackInfo().

SongInfo class was removed, replaced by
PlaybackSynchronizer::PlaybackInfo struct.
This one is not shared, every time something requests PlaybackInfo, it
is generated.

PlayerClient doesn't store PlaybackState anymore. PlaybackSynchronizer
is responsible for that.

For now, everything works just as it worked before the
changes (although now when GetPlaybackInfo() is called it adds the
time since the last update to elapsed_time and subtracts it from
remaining_time if the song is playing).

The changes are the preparation for upcomming features. Also the code
is now more readable and names are better.