Added initial crypto stuff - SPEKE protocol and BigNum class
BigNum is just a wrapper around openssl's BN_* API.

SPEKE class is an implementation of the Simple Password Exponential
Key Exchange protocol.
The implementation is based on the publication:
Feng Hao and Siamak F. Shahandashti, The SPEKE Protocol Revisited,
2014.

The contents of this commit are very much WIP. Unit testing is needed.
Until then, it's considered unsafe.