Initial work on distributing certificates
New class was added: CertExchangeServer.
Its job is to process certificate requests sent from clients and
respond with valid certificates.
The clients authenticate with a password.

Certificate has new member function: GetHash() that generates a hash
of the certificate.

SpekeSession::MessageHandler is no longer one-parameter function. It
now expects a parameter for SpekeSession reference so that the handler
can manipulate the session based on the message contents.