Add a simple authentication
The server will expect the client to use Authenticate() RPC as a first
action after connecting to prove it knows the password.
Then the server will create a session key for the client and send it
to it.
The server will expect every RPC other than Authenticate() to provide
the session key in the metadata. The key will be compared with a set
of approved session keys and access will be granted.