Fixed zkp serialization
When serializing, the program was storing the sizes of fields into the
buffer as `size_t` type. This type is not the same on all
architectures so if an another system would deserialize the zkp, it
could get confused.

This commit fixes it. It replaces `size_t` with `uint8_t`.

It's not that simple, though. The shenanigans with templating the type
of field sizes are for testing purposes.
Tests for pointer arithmetic overflow use `size_t` as a type for
`serialize()` and `deserialize()`.